feat: 서버 클라이언트의 API 호출 중 생기는 에러 객체인 HttpClientError 추가
- 기존의 upstreamError.ts 를 error.ts에 편입
